requirements.txt
gunicorn
django-cors-headers
openai
djangorestframework 
djangorestframework-simplejwt
psycopg2-binary